pyproject.toml
[tool.poetry]
name = "notifications_sprint_1"
version = "0.1.0"
description = ""
authors = [
    "Vasilii Safronov <vasi1988@mail.ru>",
    "Oleg A. <t0rr@mail.ru>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
Markdown = "^3.3.7"
python-dotenv = "^0.20.0"
psycopg2-binary = "~2.8"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"
Django = "^4.0.4"
Pillow = "^9.1.0"
loguru = "^0.6.0"
backoff = "^2.0.1"
celery = {extras = ["redis"], version = "^5.2.6"}
pika = "^1.2.1"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
django-extensions = "^3.1.5"
django-debug-toolbar = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | venv
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
multi_line_output = 3


[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = "*.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true


[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
]

ignored-modules = ["pydantic", "flask_sqlalchemy"]
good-names = []

[tool.pylint.similarities]
ignore-imports = true
ignore-docstrings = true
ignore-comments = true


[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-5.5s [%(name)s] %(message)s"
